services/guidance/requirements-dev.txt
# ──────────────────────────────────────────────
# Guidance Service — development/test dependencies
# ──────────────────────────────────────────────
-r requirements.txt

# Testing
pytest>=8.1,<9.0
pytest-cov>=5.0,<6.0
pytest-asyncio>=0.23,<1.0
hypothesis>=6.98,<7.0
moto[dynamodb,sns,sqs]>=5.0,<6.0

# Linting / type-checking
ruff>=0.3,<1.0
mypy>=1.9,<2.0
boto3-stubs[dynamodb,sns,sqs]>=1.34,<2.0

# Local dev
httpx>=0.27,<1.0
